services/requirements.txt
Django>=1.11,<1.12
validate_email
pytz
python-dateutil
six
image
crontab
beautifulsoup4
isodate
couchdb
google-api-python-client>=1.6,<1.7
celery>=4.1,<4.2
django-celery-results
coverage
MySQL-python
requests
pyzmq
djangorestframework>=3.7,<3.8
drf-extensions>=0.3,<0.4
django-ipware
Pillow>3.4,<4.0
python-memcached
sorl-thumbnail==12.3
ua-parser
user-agents
django-user-agents
selenium
django-extensions>=1.9,<1.10
xlrd
xlwt
xlsxwriter
apns
pyfcm
filelock>=2.0.6,<2.1
django-filter
markdown>=2.6.9,<2.7
google-cloud-bigquery
urllib3
lxml
future
sqlparse
django-minio-storage
django-cors-headers>=2.2,<2.3.0
django-redis